Add debug info and restore stock properly
KNOWN BUG: The only issue left is if the client asks for the whole stock
of a product, as we are solving the consequence of the bug rather than
the cause, the stock counter of the product can be below 0 and the stock
swicthes to "Out of stock". In this case, we need to set it back to "In
stock" when the stock has been restored.